+++ kamui/__init__.py
@@ -56,6 +56,7 @@ __license__ = "MIT"
 from kamui.evaluate.generation import generate
 from kamui.evaluate.perplexity import compute_perplexity
 from kamui.hooks.manager import HookManager
+from kamui.mechinterp.activation_patch import ActivationPatcher
 from kamui.mechinterp.logit_lens import LogitLens
 from kamui.model.config import ModelConfig
 from kamui.model.transformer import KAMUITransformer
@@ -66,7 +67,6 @@ from kamui.training.trainer import Trainer, TrainingConfig
 # Not yet implemented (future mechinterp tools):
 # from kamui.mechinterp.attention_viz import AttentionVisualizer
 # from kamui.mechinterp.probing import LinearProbe
-# from kamui.mechinterp.activation_patch import ActivationPatcher
 # from kamui.mechinterp.induction import InductionHeadDetector
 # from kamui.mechinterp.circuits import CircuitAblator
 
@@ -82,4 +82,5 @@ __all__: list[str] = [
     "Trainer",
     "TrainingConfig",
     "LogitLens",
+    "ActivationPatcher",
 ]